+++ collective/venue/eventaccessor.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from Products.CMFPlone.utils import safe_unicode
 from collective.address.behaviors import IAddress
 from collective.address.vocabulary import get_pycountry_name
@@ -10,6 +11,7 @@ from plone.event.interfaces import IEventAccessor
 from zope.component import adapter
 from zope.component.hooks import getSite
 from zope.interface import implementer
+from .utils import join_nonempty
 
 
 @adapter(IDXEvent)
@@ -22,8 +24,7 @@ class VenueEventAccessor(EventAccessor):
 
     @property
     def _location_link_template(self):
-        return u'<a class="venue_ref_popup" href="{url}" title="{address}">'\
-               u'{title}</a>'
+        return u'<a class="pat-plone-modal" href="{url}" title="{address}">{title}</a>'  # noqa
 
     @property
     def location(self):
@@ -31,14 +32,8 @@ class VenueEventAccessor(EventAccessor):
         location_ref = ILocation(context)
         location_uid = location_ref.location_uid
         location_notes = location_ref.location_notes
+        location_url = location_ref.location_url
         location = uuidToObject(location_uid)
-
-        site_url = getSite().absolute_url()
-        location_url = location.absolute_url()
-        if site_url not in location_url:
-            # location in different site - cannot directly open it
-            location_url = u'{0}/@@venue_view?uid={1}'.format(
-                site_url, location_uid)
 
         meta_basic = IBasic(location, None)
         add = IAddress(location, None)
@@ -47,25 +42,34 @@ class VenueEventAccessor(EventAccessor):
         if meta_basic and add:
             # I'm a location reference.
             # Create a link with href, title and urltext.
+
+            # construct url to location
+            site = getSite()
+            location_url = location.absolute_url()
+            site_path = u'/'.join(site.getPhysicalPath())
+            location_path = u'/'.join(location.getPhysicalPath())
+            if site_path not in location_path:
+                # location in different site - cannot directly open it
+                location_url = u'{0}/@@venue_view?uid={1}'.format(
+                    site.absolute_url(), location_uid)
+
             country = get_pycountry_name(add.country)
             ret = self._location_link_template.format(  # noqa
                 url=location_url,
-                address=u', '.join([it for it in [
+                address=join_nonempty((
                     add.street,
-                    add.zip_code,
-                    add.city,
+                    join_nonempty((add.zip_code, add.city), sep=u' '),
                     country
-                ] if it]),
+                ), sep=u', '),
                 title=meta_basic.title,
             )
 
         ret = safe_unicode(ret)
         location_notes = safe_unicode(location_notes)
+        if location_url:
+            location_url = u'<a href="{0}">{0}</a>'.format(location_url)
+        ret = join_nonempty([ret, location_notes, location_url], u'. ')
 
-        ret = u', '.join([it for it in [
-            safe_unicode(ret),
-            safe_unicode(location_notes)
-        ] if it])
         return ret
 
     @location.setter